+++ backend/src/test/groovy/com/example/webnovelservice/domain/payment/api/TokenTransactionApiTest.groovy
@@ -1,8 +1,8 @@
-package com.example.webnovelservice.domain.novel.api
+package com.example.webnovelservice.domain.payment.api
 
 import com.example.webnovelservice.domain.common.ApiTest
 import com.example.webnovelservice.domain.novel.steps.NovelSteps
-import com.example.webnovelservice.domain.novel.steps.TokenTransactionSteps
+import com.example.webnovelservice.domain.payment.steps.TokenTransactionSteps
 import com.example.webnovelservice.domain.user.UserRepository
 import com.example.webnovelservice.security.TokenProvider
 import com.example.webnovelservice.security.UserPrincipal
