+++ backend/src/test/groovy/com/example/webnovelservice/domain/payment/service/OwnershipTokenTransactionServiceTest.groovy
@@ -1,7 +1,11 @@
-package com.example.webnovelservice.domain.payment
+package com.example.webnovelservice.domain.payment.service
 
 import com.example.webnovelservice.domain.novel.NovelRepository
 import com.example.webnovelservice.domain.novel.entity.Novel
+import com.example.webnovelservice.domain.payment.NovelTokenCounterRepository
+import com.example.webnovelservice.domain.payment.OwnershipTokenTransactionRepository
+import com.example.webnovelservice.domain.payment.OwnershipTokenTransactionService
+import com.example.webnovelservice.domain.payment.TokenPolicyService
 import com.example.webnovelservice.domain.payment.entity.NovelTokenCounter
 import com.example.webnovelservice.domain.payment.entity.OwnershipTokenTransaction
 import com.example.webnovelservice.domain.user.UserRepository
